+++ NEWS.md
@@ -1,3 +1,9 @@
+* Added `eq()` costructor for object of class `eq`;
+* Added ui and utils functions;
+* support for meta-programming by **rlang** package;
+* Insert Logo;
+* update Travis.
+
 # equationer 0.0.0.9000
 
 * Added basic development support: